cloud_functions/requirements.txt
Flask==2.3.2
google-cloud-functions==1.1.0
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.27.0
functions-framework==3.3.0